Move RemotePlayer code to its own cpp/header
