+++ src/localplayer.h
@@ -28,6 +28,7 @@ class Client;
 class Environment;
 class GenericCAO;
 class ClientActiveObject;
+class IGameDef;
 
 enum LocalPlayerAnimations {NO_ANIM, WALK_ANIM, DIG_ANIM, WD_ANIM};  // no local animation, walking, digging, both
