+++ src/player.cpp
@@ -19,15 +19,12 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 #include "player.h"
 
-#include <fstream>
 #include "threading/mutex_auto_lock.h"
 #include "util/numeric.h"
 #include "hud.h"
 #include "constants.h"
 #include "gamedef.h"
 #include "settings.h"
-#include "content_sao.h"
-#include "filesys.h"
 #include "log.h"
 #include "porting.h"  // strlcpy
 
@@ -143,206 +140,3 @@ void Player::clearHud()
 		hud.pop_back();
 	}
 }
-
-/*
-	RemotePlayer
-*/
-// static config cache for remoteplayer
-bool RemotePlayer::m_setting_cache_loaded = false;
-float RemotePlayer::m_setting_chat_message_limit_per_10sec = 0.0f;
-u16 RemotePlayer::m_setting_chat_message_limit_trigger_kick = 0;
-
-RemotePlayer::RemotePlayer(const char *name, IItemDefManager *idef):
-	Player(name, idef),
-	protocol_version(0),
-	m_sao(NULL),
-	m_dirty(false),
-	m_last_chat_message_sent(time(NULL)),
-	m_chat_message_allowance(5.0f),
-	m_message_rate_overhead(0),
-	hud_hotbar_image(""),
-	hud_hotbar_selected_image("")
-{
-	if (!RemotePlayer::m_setting_cache_loaded) {
-		RemotePlayer::m_setting_chat_message_limit_per_10sec =
-				g_settings->getFloat("chat_message_limit_per_10sec");
-		RemotePlayer::m_setting_chat_message_limit_trigger_kick =
-				g_settings->getU16("chat_message_limit_trigger_kick");
-		RemotePlayer::m_setting_cache_loaded = true;
-	}
-	movement_acceleration_default   = g_settings->getFloat("movement_acceleration_default")   * BS;
-	movement_acceleration_air       = g_settings->getFloat("movement_acceleration_air")       * BS;
-	movement_acceleration_fast      = g_settings->getFloat("movement_acceleration_fast")      * BS;
-	movement_speed_walk             = g_settings->getFloat("movement_speed_walk")             * BS;
-	movement_speed_crouch           = g_settings->getFloat("movement_speed_crouch")           * BS;
-	movement_speed_fast             = g_settings->getFloat("movement_speed_fast")             * BS;
-	movement_speed_climb            = g_settings->getFloat("movement_speed_climb")            * BS;
-	movement_speed_jump             = g_settings->getFloat("movement_speed_jump")             * BS;
-	movement_liquid_fluidity        = g_settings->getFloat("movement_liquid_fluidity")        * BS;
-	movement_liquid_fluidity_smooth = g_settings->getFloat("movement_liquid_fluidity_smooth") * BS;
-	movement_liquid_sink            = g_settings->getFloat("movement_liquid_sink")            * BS;
-	movement_gravity                = g_settings->getFloat("movement_gravity")                * BS;
-}
-
-void RemotePlayer::save(std::string savedir, IGameDef *gamedef)
-{
-	/*
-	 * We have to open all possible player files in the players directory
-	 * and check their player names because some file systems are not
-	 * case-sensitive and player names are case-sensitive.
-	 */
-
-	// A player to deserialize files into to check their names
-	RemotePlayer testplayer("", gamedef->idef());
-
-	savedir += DIR_DELIM;
-	std::string path = savedir + m_name;
-	for (u32 i = 0; i < PLAYER_FILE_ALTERNATE_TRIES; i++) {
-		if (!fs::PathExists(path)) {
-			// Open file and serialize
-			std::ostringstream ss(std::ios_base::binary);
-			serialize(ss);
-			if (!fs::safeWriteToFile(path, ss.str())) {
-				infostream << "Failed to write " << path << std::endl;
-			}
-			setModified(false);
-			return;
-		}
-		// Open file and deserialize
-		std::ifstream is(path.c_str(), std::ios_base::binary);
-		if (!is.good()) {
-			infostream << "Failed to open " << path << std::endl;
-			return;
-		}
-		testplayer.deSerialize(is, path);
-		is.close();
-		if (strcmp(testplayer.getName(), m_name) == 0) {
-			// Open file and serialize
-			std::ostringstream ss(std::ios_base::binary);
-			serialize(ss);
-			if (!fs::safeWriteToFile(path, ss.str())) {
-				infostream << "Failed to write " << path << std::endl;
-			}
-			setModified(false);
-			return;
-		}
-		path = savedir + m_name + itos(i);
-	}
-
-	infostream << "Didn't find free file for player " << m_name << std::endl;
-	return;
-}
-
-void RemotePlayer::deSerialize(std::istream &is, const std::string &playername)
-{
-	Settings args;
-
-	if (!args.parseConfigLines(is, "PlayerArgsEnd")) {
-		throw SerializationError("PlayerArgsEnd of player " +
-								 playername + " not found!");
-	}
-
-	m_dirty = true;
-	//args.getS32("version"); // Version field value not used
-	std::string name = args.get("name");
-	strlcpy(m_name, name.c_str(), PLAYERNAME_SIZE);
-	setPitch(args.getFloat("pitch"));
-	setYaw(args.getFloat("yaw"));
-	setPosition(args.getV3F("position"));
-	try{
-		hp = args.getS32("hp");
-	}catch(SettingNotFoundException &e) {
-		hp = PLAYER_MAX_HP;
-	}
-	try{
-		m_breath = args.getS32("breath");
-	}catch(SettingNotFoundException &e) {
-		m_breath = PLAYER_MAX_BREATH;
-	}
-
-	inventory.deSerialize(is);
-
-	if(inventory.getList("craftpreview") == NULL) {
-		// Convert players without craftpreview
-		inventory.addList("craftpreview", 1);
-
-		bool craftresult_is_preview = true;
-		if(args.exists("craftresult_is_preview"))
-			craftresult_is_preview = args.getBool("craftresult_is_preview");
-		if(craftresult_is_preview)
-		{
-			// Clear craftresult
-			inventory.getList("craftresult")->changeItem(0, ItemStack());
-		}
-	}
-}
-
-void RemotePlayer::serialize(std::ostream &os)
-{
-	// Utilize a Settings object for storing values
-	Settings args;
-	args.setS32("version", 1);
-	args.set("name", m_name);
-	//args.set("password", m_password);
-	args.setFloat("pitch", m_pitch);
-	args.setFloat("yaw", m_yaw);
-	args.setV3F("position", m_position);
-	args.setS32("hp", hp);
-	args.setS32("breath", m_breath);
-
-	args.writeLines(os);
-
-	os<<"PlayerArgsEnd\n";
-
-	inventory.serialize(os);
-}
-
-void RemotePlayer::setPosition(const v3f &position)
-{
-	if (position != m_position)
-		m_dirty = true;
-
-	Player::setPosition(position);
-	if(m_sao)
-		m_sao->setBasePosition(position);
-}
-
-const RemotePlayerChatResult RemotePlayer::canSendChatMessage()
-{
-	// Rate limit messages
-	u32 now = time(NULL);
-	float time_passed = now - m_last_chat_message_sent;
-	m_last_chat_message_sent = now;
-
-	// If this feature is disabled
-	if (m_setting_chat_message_limit_per_10sec <= 0.0) {
-		return RPLAYER_CHATRESULT_OK;
-	}
-
-	m_chat_message_allowance += time_passed * (m_setting_chat_message_limit_per_10sec / 8.0f);
-	if (m_chat_message_allowance > m_setting_chat_message_limit_per_10sec) {
-		m_chat_message_allowance = m_setting_chat_message_limit_per_10sec;
-	}
-
-	if (m_chat_message_allowance < 1.0f) {
-		infostream << "Player " << m_name
-				<< " chat limited due to excessive message amount." << std::endl;
-
-		// Kick player if flooding is too intensive
-		m_message_rate_overhead++;
-		if (m_message_rate_overhead > RemotePlayer::m_setting_chat_message_limit_trigger_kick) {
-			return RPLAYER_CHATRESULT_KICK;
-		}
-
-		return RPLAYER_CHATRESULT_FLOODING;
-	}
-
-	// Reinit message overhead
-	if (m_message_rate_overhead > 0) {
-		m_message_rate_overhead = 0;
-	}
-
-	m_chat_message_allowance -= 1.0f;
-	return RPLAYER_CHATRESULT_OK;
-}
-
