+++ src/script/lua_api/l_object.h
@@ -47,7 +47,7 @@ private:
 
 	static PlayerSAO* getplayersao(ObjectRef *ref);
 
-	static RemotePlayer* getplayer(ObjectRef *ref);
+	static RemotePlayer *getplayer(ObjectRef *ref);
 
 	// Exported functions
